server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
anthropic==0.28.0
google-generativeai==0.5.4
beautifulsoup4==4.12.3
markdownify==0.12.1
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
